MR-262: fix issues and update folders
