+++ src/Components/Member/Profile/index.jsx
@@ -6,7 +6,7 @@ import { joiResolver } from '@hookform/resolvers/joi';
 import styles from './profile.module.css';
 import Button from 'Components/Shared/Button';
 import Input from 'Components/Shared/Input/index';
-import memberValidation from 'Validations/members';
+import memberValidation from 'Validations/Admin/members';
 import Container from 'Components/Shared/Container';
 import toast, { Toaster } from 'react-hot-toast';
 import { updateUser } from 'Redux/auth/actions';
