+++ src/Components/Shared/Header/DropDownMenu/dropDownMenu.moule.css
@@ -19,7 +19,8 @@
 
 .dropdown-menu {
   position: absolute;
-  top: 14vh;
+  top: 11.5vh;
+  box-shadow: 0px 6px 7px -3.5px rgba(0,0,0,0.2), 0px 12px 19px 1.5px rgba(0,0,0,0.14), 0px 4.5px 23px 4px rgba(0,0,0,0.12);
   right: 7.3vh;
   background-color: #f3f3f3;
   border-radius: var(--border-radius);
