+++ src/Redux/auth/constants.js
@@ -7,7 +7,7 @@ export const LOGIN_ERROR = 'LOGIN_ERROR';
 export const LOGOUT_PENDING = 'LOGOUT_PENDING';
 export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
 export const LOGOUT_ERROR = 'LOGOUT_ERROR';
-export const SET_USER_ROLE = 'SET_USER_ROLE';
+export const SET_USER_AUTH = 'SET_USER_AUTH';
 export const UPDATE_USER = 'UPDATE_USER';
 export const SIGN_UP_PENDING = 'SIGN_UP_PENDING';
 export const SIGN_UP_SUCCESS = 'SIGN_UP_SUCCESS';
